requirements.txt
streamlit==1.28.0
pandas==2.1.0
numpy==1.24.0
pdfplumber==0.10.2
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow==10.0.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
scikit-learn==1.3.0
requests==2.31.0
openai==1.3.0
anthropic==0.7.1
google-generativeai==0.3.0
python-dotenv==1.0.0
pydantic==2.4.0